Refactor in source code to export interfaces in API
