+++ dynamiclibrary/Observer.h
@@ -1,0 +1,25 @@
+#pragma once
+
+#include "pch.h"
+
+namespace ObserverApi
+{
+    class Observer : public IObserver
+    {
+    public:
+        Observer(int id_number, Subject& subject);
+
+        virtual ~Observer();
+
+        void Update(const std::string& message_from_subject);
+
+        void RemoveMeFromTheList();
+
+        void PrintInfo();
+
+    private:
+        std::string message_from_subject_;
+        Subject& subject_;
+        int _id_number = 0;
+    };
+}
