+++ dynamiclibrary/dynamiclibrary.vcxproj.filters
@@ -27,13 +27,16 @@
     <ClInclude Include="pch.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="ObserverApi.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="IDynamicLibraryApi.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="Operation.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="Subject.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="Observer.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
