+++ dynamiclibrary/pch.h
@@ -16,7 +16,9 @@
 // add headers that you want to pre-compile here
 #include "framework.h"
 #include "IDynamicLibraryApi.h"
-#include "ObserverApi.h"
+//#include "ObserverApi.h"
+#include "Subject.h"
+#include "Observer.h"
 #include "Operation.h"
 
 #endif //PCH_H
